Switch from sqlite to postgresql

+++ db/migrate/20240229145453_remove_otp_fields_from_users.rb
@@ -1,7 +1,9 @@
 class RemoveOtpFieldsFromUsers < ActiveRecord::Migration[7.1]
   def change
-    remove_column :users, :otp_secret, :string
-    remove_column :users, :consumed_timestep, :integer
-    remove_column :users, :otp_required_for_login, :boolean
+    remove_column :users, :otp_secret, :string if column_exists?(:users, :otp_secret)
+
+    remove_column :users, :consumed_timestep, :integer if column_exists?(:users, :consumed_timestep)
+
+    remove_column :users, :otp_required_for_login, :boolean if column_exists?(:users, :otp_required_for_login)
   end
 end
